support qwen3 moe liger, split registration of liger into auto.py
