+++ veomni/models/transformers/qwen3_moe/qwen3_moe_gpu_patch_gen_config.py
@@ -23,8 +23,6 @@ This keeps only the needed v5 patches:
 3. Register get_parallel_plan on Qwen3MoeForCausalLM.
 """
 
-from typing import Optional
-
 import torch
 from transformers.activations import ACT2FN
 from transformers.cache_utils import Cache, DynamicCache
@@ -35,7 +33,7 @@ from transformers.processing_utils import Unpack
 from transformers.utils import TransformersKwargs
 
 from veomni.ops import fused_moe_forward
-from veomni.patchgen.patch_spec import PatchConfig, create_patch_from_external
+from veomni.patchgen.patch_spec import PatchConfig
 
 
 config = PatchConfig(
@@ -46,23 +44,23 @@ config = PatchConfig(
 
 config.add_import("veomni.ops", names=["fused_moe_forward"])
 
-config.patches.append(
-    create_patch_from_external(
-        target="Qwen3MoeRMSNorm",
-        replacement_module="liger_kernel.transformers.rms_norm",
-        replacement_name="LigerRMSNorm",
-        description="Use LigerKernel RMSNorm",
-    )
-)
+# config.patches.append(
+#     create_patch_from_external(
+#         target="Qwen3MoeRMSNorm",
+#         replacement_module="liger_kernel.transformers.rms_norm",
+#         replacement_name="LigerRMSNorm",
+#         description="Use LigerKernel RMSNorm",
+#     )
+# )
 
-config.patches.append(
-    create_patch_from_external(
-        target="Qwen3MoeMLP",
-        replacement_module="liger_kernel.transformers.swiglu",
-        replacement_name="LigerSwiGLUMLP",
-        description="Use LigerKernel SwiGLU MLP",
-    )
-)
+# config.patches.append(
+#     create_patch_from_external(
+#         target="Qwen3MoeMLP",
+#         replacement_module="liger_kernel.transformers.swiglu",
+#         replacement_name="LigerSwiGLUMLP",
+#         description="Use LigerKernel SwiGLU MLP",
+#     )
+# )
 
 
 @config.replace_class(
@@ -126,25 +124,25 @@ class PatchedQwen3MoeExperts(torch.nn.Module):
         return final_hidden_states
 
 
-@config.replace_function("apply_rotary_pos_emb", description="Use LigerKernel rotary embedding")
-def apply_rotary_pos_emb_liger(
-    q: torch.Tensor,
-    k: torch.Tensor,
-    cos: torch.Tensor,
-    sin: torch.Tensor,
-    position_ids: Optional[torch.Tensor] = None,
-    unsqueeze_dim: int = 1,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    from liger_kernel.transformers.rope import liger_rotary_pos_emb
+# @config.replace_function("apply_rotary_pos_emb", description="Use LigerKernel rotary embedding")
+# def apply_rotary_pos_emb_liger(
+#     q: torch.Tensor,
+#     k: torch.Tensor,
+#     cos: torch.Tensor,
+#     sin: torch.Tensor,
+#     position_ids: Optional[torch.Tensor] = None,
+#     unsqueeze_dim: int = 1,
+# ) -> tuple[torch.Tensor, torch.Tensor]:
+#     from liger_kernel.transformers.rope import liger_rotary_pos_emb
 
-    return liger_rotary_pos_emb(
-        q,
-        k,
-        cos,
-        sin,
-        position_ids=position_ids,
-        unsqueeze_dim=unsqueeze_dim,
-    )
+#     return liger_rotary_pos_emb(
+#         q,
+#         k,
+#         cos,
+#         sin,
+#         position_ids=position_ids,
+#         unsqueeze_dim=unsqueeze_dim,
+#     )
 
 
 @config.override_method("Qwen3MoeModel.forward", description="Support SP in Qwen3MoeModel.forward")
